Add a standard but generic document resolution
In the tests, adds a simple provides for a terse but precise syntax
intended to be used for Table of Contents or similar files. In
particular, we show that `//TOC` can be served from a different domain
than other files. This solves ambiguities with standard file references
but adds the benefit of extensibility.