Add a clean front page to reports
The cover is similar to thesis again, but without supervisor, advisor
and additional overhead. Most items should be optional.

This reuses the definitions made by the thesis target, thus moving them
to a common module. Note that the `ifempty` function appeared broken but
is more important for a simple report.