Allow inline latex commands again
b7250c7ee51d18f952658bc6771354d727bf9536 introduced
converting `\` with `\textbackslash{}` within text in
the latex backend.
This removes the ability of using inline-latex commands.
This commit introduces a rough regex which tests if something
following a `\` in text could be latex command, in which case
it is not escaped.