Expect failed canonicalization correctly
Canonicalization is required to ensure proper access rights and failure
to do so (for example from a missing file) should not be treated as an
internal implementation failure. Rather, it is most likely the input
file's fault from a malformed path.

An intermediate error type diagnoses different problems with path
normalization when resolving a relative file.