Fix thesis to use an absolute context
This means you can't use relative links in it but that's really somewhat
caused by the resolution of the abstracts to happen within the
configuration frontend instead of the generator. There is also an
overlap where latex specific environment variables are set depending the
template and that should likely only happen in the latex backend as
well.