Restructure math equation block generators
The end event is necessarily fits the beginning one as the causing front
end event is the end of the corresponding Markdown code block. This
makes redundant to check for this correspondence by splitting the tags.
It also wrongly suggests duplicating the logic for handling the (latex
formatted) math content.