Introduce new backend specializing beamer
This required significantly less work than imagined. Also, it prepares
open and close calls for customization as we'll want to be notified of
every opened frame to keep track of the index for synchronizing audio to
the frames which we will later extract individually from the built pdf.
This type parameter _somehow_ defaults to `()` when it is not provided
which means there is no need to change _any_ other site of use by simply
adding the respective impls for the unit tuple as well.