Include doc comments as markdown
Uses the header adjustment strategy previously introduced to the code to
ensure that scoping is correct. Most doc comments contain an h1 instead
of the proper header level from the context. Note that in particular
this will put some pressure on the backend to allow for slightly deeper
nested headers, at least to 6. We already emulate them by a weighted
font in one case.